Move exception to commons/exceptions
